Добавлены базовые роуты
