+++ package.json
@@ -13,6 +13,7 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.2",
     "react-scripts": "^2.1.3",
     "web-vitals": "^2.1.4"
   },
