+++ src/index.js
@@ -1,6 +1,7 @@
 // import { ThemeProvider } from '@mui/styles';
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { BrowserRouter } from 'react-router-dom';
 import App from './App';
 // import theme from './theme';
 // import { createTheme } from '@mui/material/styles';
@@ -17,7 +18,9 @@ import App from './App';
 
 
 ReactDOM.render(
-  <App />,
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>,
   document.getElementById('root')
 );
