+++ src/pages/Home.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Home(props) {
+    return (
+        <div sx="color: black">
+            <h1>Home Page</h1>
+        </div>
+    );
+}
+
+export default Home;
