+++ src/pages/NoFoundPage.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function NoFoundPage(props) {
+    return (
+        <div>
+            <h1>Page Not Found</h1>
+        </div>
+    );
+}
+
+export default NoFoundPage;
