+++ src/pages/Profile.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Profile(props) {
+    return (
+        <div>
+            <h1>Profile Page</h1>
+        </div>
+    );
+}
+
+export default Profile;
